+++ backend/src/main/java/com/devu/backend/api/recruit/RecruitController.java
@@ -1,0 +1,14 @@
+package com.devu.backend.api.recruit;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api")
+public class RecruitController {
+    //TODO Recruit 반환 api
+}
