+++ backend/src/main/resources/application-local.yml
@@ -20,3 +20,7 @@ spring:
         dialect: org.hibernate.dialect.H2Dialect
     hibernate:
       ddl-auto: update
+
+webDriverId: webdriver.chrome.driver
+webDriverPath: C:/chromedriver.exe
+
